[refactor] dedupe attention-mask helpers in llama3_eagle
llama3_eagle.py carried byte-identical copies of _make_causal_mask and
_expand_mask from modeling/_mask_utils.py, plus a module-level
prepare_decoder_attention_mask duplicating the Eagle3DraftModel method.

Move prepare_decoder_attention_mask into _mask_utils.py as the single
home, have the base-class method delegate to it, drop all three copies
from llama3_eagle.py, and call self.prepare_decoder_attention_mask in
LlamaForCausalLMEagle3.forward. The benchmark and attention tests that
imported the module-level helper now import it from _mask_utils.
